api/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
sqlalchemy==2.0
psycopg2==2.9.11
passlib==1.7.4
pyjwt==2.10.1
python-multipart==0.0.20